+++ kernel/include/tests.h
@@ -9,4 +9,8 @@ void test_heap(void);
 void test_vam(void);
 void test_vmm(void);
 
+void test_process(void);
+void test_thread(void);
+void test_scheduler(void);
+
 #endif